Remove Heroku section
I don't think having a section like this in the project adds any value on a day-to-day basis.
It does however add the burden of keeping the prices update, and the actual heroku setup in sync with what's in the README.

I understand if the point of the section was to guide developers on how to set up heroku when initially creating the project. In that case let's add such a section to the template's README instead.