Refactor: Move valid_token? to user model.
Actually the ApiAuthenticable module which the user model includes.
Since this is where the token generation is, it makes sense to have the
validation there too.

One thing I don't like about this now is that the token validation is
now tied to the http header convention of prefixing the token with
"Token ", which is not strictly necessary for raw token comparison.

But since the module is ApiAuthenticable it's probably a safe bet we'll
always want to be comparing http headers, so I'm not going to do
anything about "fixing" that yet.