Remove PROJECT_DOMAIN.
It is better to set it in an environment variable than save it in the repo.
It is still possible to save it in the repo, if wanted, by setting it in .env.
This way supports both wishes.

Also, it isn't necessary to specify domain for the smtp_settings.
From Rails mailer guide:
> :domain - If you need to specify a HELO domain, you can do it here.

The HELO domain might be added as a Received mail header. So if we add it, it should be a domain that identifies the webserver.
https://en.wikipedia.org/wiki/Simple_Mail_Transfer_Protocol
> The information that the client sends in the HELO and MAIL FROM commands are added (not seen in example code) as additional header fields to the message by the receiving server. It adds a Received and Return-Path header field, respectively.