CI: comment with rspec-stderr (#2330)
We should avoid introducing noisy warnings in our test output.
Often the warnings (deprecation warnings) are added by Dependabot upgrades. This way it will not merge PR adding noise automatically.

https://app.asana.com/0/1142794766483633/1208315235207762

`continue-on-error` because we do not want this marked as the primary failure (scrolled to and opened), if any other steps failed.
It will still output the warnings as github actions markup errors, so they will be visible elsewhere.